requirements.txt
azure-core==1.35.0
azure-eventhub==5.15.0
azure-functions==1.23.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
MarkupSafe==3.0.2
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
